pyproject.toml
[project]
name = "cripto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "binance-sdk-wallet>=1.2.0",
    "python-dotenv>=1.1.1",
]
